requirements.txt
# IoT Smart Parking System - Backend Dependencies

# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0

# MQTT and Communication
paho-mqtt==1.6.1
requests==2.31.0

# Caching and Session Management
redis==5.0.0
Flask-Session==0.5.0

# Payment Processing
stripe==6.6.0

# Data Processing and Analytics
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Logging and Monitoring
structlog==23.1.0

# WebSocket support
eventlet==0.33.3

# API Documentation
flask-restx==1.2.0

# Security
bcrypt==4.0.1
cryptography==41.0.4